Movendo método de validação para pacote utils
